Update New Theatre Square.cpp

+++ 1000/New Theatre Square.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-	int t;
+	ll t;
 	cin>>t;
 	while(t--)
 	{
